requirements.txt
async-timeout==4.0.3
bleak==0.21.1
bleak-winrt==1.2.0
install==1.3.5
PyYAML==6.0.1
typing_extensions==4.9.0
